PyBank // PyPoll
Went back to the beginning of PyBank to create lists and variables that were necessary to properly calculate the Profit/Loss from each month. Originally I didn't realize that the column as a whole needed to be recalculated.    For PyPoll, I retooled the variables/lists to make it easier to determine the total votes and percentages for rach candidate. Then,  conditionals were placed to append the total votes into candidate-specific lists and to determine the overall winner.