PyPoll
Completed comments in code to help others read through easily and understand my thinking and workflow. Could not find a way to get the li_percent to format properly. I attempted running the variable through the debugger, rounding the variable earlier up in the script. and creating a new variable that already had the rounded value and none of the three remedies worked. Frustrating, but c'est la vie!